+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
             android:name=".feature.orderDetail.OrderDetailActivity"
             android:exported="false" />
         <activity
-            android:name=".feature.order.OrderActivity"
+            android:name=".feature.userInfo.UserInfoActivity"
             android:exported="false" />
         <activity
             android:name=".feature.payment.PaymentActivity"
