+++ app/src/main/java/woowacourse/shopping/feature/userInfo/UserInfoContract.kt
@@ -1,9 +1,9 @@
-package woowacourse.shopping.feature.order
+package woowacourse.shopping.feature.userInfo
 
 import woowacourse.shopping.model.OrderUiModel
 import woowacourse.shopping.model.PointUiModel
 
-interface OrderContract {
+interface UserInfoContract {
     interface View {
         fun showOrders(orders: List<OrderUiModel>)
         fun showPoint(point: PointUiModel)
