+++ app/src/main/res/values/strings.xml
@@ -35,4 +35,5 @@
     <string name="order_history">주문 내역</string>
     <string name="payment_info">결제 정보</string>
     <string name="order_at">주문 일시</string>
+    <string name="user_page">User Page</string>
 </resources>
